Reset q-values (#52)
* Reset q-values 

Reset q-values between conditions in hddm.HDDMrl and hddm.Hrl

* Update rl.py

prevent type-error for feedback-data

* Update hddm_rl.py

prevent type-error for feedback and split_by-data

* Update rl.py

ensure response is int

+++ hddm/models/rl.py
@@ -79,7 +79,7 @@ def RL_like(x, v, alpha, pos_alpha, z=0.5, p_outlier=0):
     wp = wiener_params
     response = x['response'].values.astype(int)
     q = x['q_init'].iloc[0]
-    feedback = x['feedback'].values
-    split_by = x['split_by'].values
+    feedback = x['feedback'].values.astype(float)
+    split_by = x['split_by'].values.astype(int)
     return wiener_like_rl(response, feedback, split_by, q, alpha, pos_alpha, v, z, p_outlier=p_outlier, **wp)
 RL = stochastic_from_dist('RL', RL_like)
